Move DocumentsUtils to common package, implement catalog process add interface by template
