+++ packages/backend-for-frontend/src/model/errors.ts
@@ -60,6 +60,7 @@ export const errorCodes = {
   tooManyDescriptorForInterfaceWithTemplate: "0051",
   eserviceDescriptorDraftNotFound: "0052",
   templateDataNotFound: "0053",
+  noVersionInEServiceTemplate: "0045",
 };
 
 export type ErrorCodes = keyof typeof errorCodes;
@@ -308,34 +309,6 @@ export function invalidInterfaceContentTypeDetected(
   });
 }
 
-export function invalidInterfaceFileDetected(
-  eServiceId: string
-): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `The interface file for EService ${eServiceId} is invalid`,
-    code: "invalidInterfaceFileDetected",
-    title: "Invalid interface file detected",
-  });
-}
-
-export function openapiVersionNotRecognized(
-  version: string
-): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `OpenAPI version not recognized - ${version}`,
-    code: "openapiVersionNotRecognized",
-    title: "OpenAPI version not recognized",
-  });
-}
-
-export function interfaceExtractingInfoError(): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `Error extracting info from interface file`,
-    code: "interfaceExtractingInfoError",
-    title: "Error extracting info from interface file",
-  });
-}
-
 export function notValidDescriptor(
   descriptorId: string,
   state: string
@@ -479,16 +452,6 @@ export function eserviceTemplateNotFound(
   });
 }
 
-export function eserviceTemplateDataNotFound(
-  eServiceId: string
-): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `Expected template reference in Eservice ${eServiceId} not found`,
-    code: "templateDataNotFound",
-    title: "EService template not found",
-  });
-}
-
 export function eserviceIsNotDraft(eserviceId: string): ApiError<ErrorCodes> {
   return new ApiError({
     detail: `EService ${eserviceId} is not in draft state`,
@@ -504,34 +467,5 @@ export function eserviceTemplateNotPublished(
     detail: `EService template ${eserviceTemplateId} is not in published state`,
     code: "eserviceTemplateIsNotPublished",
     title: "EService template is not in published state",
-  });
-}
-
-export function eserviceTemplateInterfaceNotFound(
-  eserviceTemplateId: string,
-  eserviceTemplateVersionId: string
-): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `EService template interface for template ${eserviceTemplateId} with version ${eserviceTemplateVersionId} not found`,
-    code: "eserviceTemplateInterfaceNotFound",
-    title: "EService template interface document not found",
-  });
-}
-
-export function eserviceInterfaceDataNotValid(): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `EService template interface data not valid`,
-    code: "eserviceTemplateInterfaceDataNotValid",
-    title: "EService template interface data not valid",
-  });
-}
-
-export function eserviceDescriptorDraftNotFound(
-  eserviceId: string
-): ApiError<ErrorCodes> {
-  return new ApiError({
-    detail: `Draft descriptor not found in Eservice ${eserviceId}`,
-    code: "eserviceDescriptorDraftNotFound",
-    title: "EService descriptor draft not found",
   });
 }
