+++ packages/backend-for-frontend/src/routers/catalogRouter.ts
@@ -62,9 +62,7 @@ const catalogRouter = (
     tenantProcessClient,
     attributeProcessClient,
     catalogProcessClient,
-    delegationProcessClient,
-    fileManager,
-    config
+    fileManager
   );
 
   catalogRouter
