+++ packages/catalog-process/.env
@@ -29,3 +29,6 @@ PRODUCER_ALLOWED_ORIGINS="IPA"
 
 FEATURE_FLAG_SIGNALHUB_WHITELIST=true
 SIGNALHUB_WHITELIST_PRODUCER=cec6866f-1561-4c83-a230-c91b98c2a890
+
+ESERVICE_TEMPLATE_DOCUMENTS_CONTAINER="interop-local-bucket"
+ESERVICE_TEMPLATE_DOCUMENTS_PATH="interop-eservice-template-documents"
