+++ packages/catalog-process/src/config/config.ts
@@ -8,6 +8,17 @@ import {
 } from "pagopa-interop-commons";
 import { z } from "zod";
 
+export const EServiceTemplateS3Config = z
+  .object({
+    ESERVICE_TEMPLATE_DOCUMENTS_CONTAINER: z.string(),
+    ESERVICE_TEMPLATE_DOCUMENTS_PATH: z.string(),
+  })
+  .transform((c) => ({
+    eserviceTemplateDocumentsContainer: c.ESERVICE_TEMPLATE_DOCUMENTS_CONTAINER,
+    eserviceTemplateDocumentsPath: c.ESERVICE_TEMPLATE_DOCUMENTS_PATH,
+  }));
+export type EServiceTemplateS3Config = z.infer<typeof EServiceTemplateS3Config>;
+
 const CatalogProcessConfig = CommonHTTPServiceConfig.and(ReadModelDbConfig)
   .and(FileManagerConfig)
   .and(S3Config)
@@ -25,7 +36,8 @@ const CatalogProcessConfig = CommonHTTPServiceConfig.and(ReadModelDbConfig)
           .map((origin) => origin.trim())
           .filter(Boolean),
       }))
-  );
+  )
+  .and(EServiceTemplateS3Config);
 
 export type CatalogProcessConfig = z.infer<typeof CatalogProcessConfig>;
