+++ packages/commons/package.json
@@ -20,21 +20,26 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
+    "@apidevtools/swagger-parser": "10.1.0",
+    "@types/adm-zip": "0.5.5",
     "@aws-sdk/client-kms": "3.693.0",
     "@aws-sdk/client-s3": "3.693.0",
     "@aws-sdk/client-sesv2": "3.693.0",
     "@aws-sdk/s3-request-presigner": "3.693.0",
     "@zodios/core": "10.9.6",
     "@zodios/express": "10.6.1",
+    "adm-zip": "0.5.15",
     "axios": "1.7.4",
     "connection-string": "4.4.0",
     "date-fns": "3.6.0",
     "date-fns-tz": "3.1.3",
     "express": "4.20.0",
+    "fast-xml-parser": "4.4.1",
     "handlebars": "4.7.8",
     "jsonwebtoken": "9.0.2",
     "jwks-rsa": "3.1.0",
     "kafkajs": "2.2.4",
+    "mime": "4.0.4",
     "mongodb": "6.7.0",
     "nodemailer": "6.9.14",
     "pagopa-interop-models": "workspace:*",
@@ -44,6 +49,7 @@
     "redis": "4.6.15",
     "ts-pattern": "5.2.0",
     "winston": "3.13.0",
+    "yaml": "2.4.5",
     "zod": "3.23.8",
     "zod-validation-error": "3.3.0"
   },
