+++ packages/commons/src/index.ts
@@ -22,8 +22,10 @@ export * from "./utils/getAll.js";
 export * from "./utils/apiUtils.js";
 export * from "./utils/date.js";
 export * from "./utils/tenantMails.js";
+export * from "./utils/fileUtils.js";
 export * from "./utils/job.js";
 export * from "./auth/jwk.js";
 export * from "./auth/converters.js";
 export * from "./email-manager/index.js";
 export * from "./rate-limiter/index.js";
+export * from "./eservice-documents/index.js";
