+++ packages/commons/src/utils/fileUtils.ts
@@ -1,0 +1,20 @@
+import { Readable } from "stream";
+import crypto from "crypto";
+
+export async function calculateChecksum(stream: Readable): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const hash = crypto.createHash("sha256");
+
+    stream.on("data", (data) => {
+      hash.update(data);
+    });
+
+    stream.on("end", () => {
+      resolve(hash.digest("hex"));
+    });
+
+    stream.on("error", (err) => {
+      reject(err);
+    });
+  });
+}
